// src/app/models/software-catalog.d.ts
export interface components {
    schemas: {
// ...
        Characteristic: {
            /** @description Unique identifier of the characteristic */
            id?: string;
            /** @description Name of the characteristic */
            name: string;
            /** @description Data type of the value of the characteristic */
            valueType?: string;
            characteristicRelationship?: components["schemas"]["CharacteristicRelationship"][];
            value: components["schemas"]["Any"];
            /** @description When sub-classing, this defines the super-class */
            "@baseType"?: string;
            /**
             * Format: uri
             * @description A URI to a JSON-Schema file that defines additional attributes and relationships
             */
            "@schemaLocation"?: string;
            /** @description When sub-classing, this defines the sub-class Extensible name */
            "@type"?: string;
        };
// ...
        Resource_Create: {
            /** @description Category of the concrete resource. e.g Gold, Silver for MSISDN concrete resource */
            category?: string;
            /** @description free-text description of the resource */
            description?: string;
            /**
             * Format: date-time
             * @description A date time( DateTime). The date till the resource is operating
             */
            endOperatingDate?: string;
            /** @description A string used to give a name to the resource */
            name: string;
            /** @description A field that identifies the specific version of an instance of a resource. */
            resourceVersion?: string;
            /**
             * Format: date-time
             * @description A date time( DateTime). The date from which the resource is operating
             */
            startOperatingDate?: string;
            /** @description Configuration features */
            activationFeature?: components["schemas"]["Feature"][];
            administrativeState?: components["schemas"]["ResourceAdministrativeStateType"];
            attachment?: components["schemas"]["AttachmentRefOrValue"][];
            note?: components["schemas"]["Note"][];
            operationalState?: components["schemas"]["ResourceOperationalStateType"];
            place?: components["schemas"]["RelatedPlaceRefOrValue"];
            relatedParty?: components["schemas"]["RelatedParty"][];
            resourceCharacteristic?: components["schemas"]["Characteristic"][];
            resourceRelationship?: components["schemas"]["ResourceRelationship"][];
            resourceSpecification?: components["schemas"]["ResourceSpecificationRef"];
            resourceStatus?: components["schemas"]["ResourceStatusType"];
            usageState?: components["schemas"]["ResourceUsageStateType"];
            /** @description When sub-classing, this defines the super-class */
            "@baseType"?: string;
            /**
             * Format: uri
             * @description A URI to a JSON-Schema file that defines additional attributes and relationships
             */
            "@schemaLocation"?: string;
            /** @description When sub-classing, this defines the sub-class Extensible name */
            "@type": string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}